refactor ampAds check
