+++ src/app/containers/Ad/index.jsx
@@ -12,9 +12,8 @@ const AdContainer = ({ slotType }) => {
   const { ads } = useContext(ServiceContext);
   const hasAds = pathOr(false, ['hasAds'], ads);
   const { enabled: adsEnabled } = useToggle('ads');
-  const { enabled: ampAdsEnabled } = useToggle('ampAds');
 
-  if (!hasAds || !adsEnabled || (isAmp && !ampAdsEnabled)) {
+  if (!hasAds || !adsEnabled) {
     return null;
   }
