+++ src/app/containers/Ad/index.test.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import { RequestContextProvider } from '#contexts/RequestContext';
@@ -122,68 +121,6 @@ describe('Ad Container', () => {
           </RequestContextProvider>
         </ServiceContextProvider>,
       );
-    });
-  });
-
-  describe('Assertions', () => {
-    const defaultToggleState = {
-      ads: {
-        enabled: true,
-      },
-      ampAds: {
-        enabled: false,
-      },
-    };
-
-    const mockToggleDispatch = jest.fn();
-
-    const toggleContextMock = {
-      toggleState: defaultToggleState,
-      toggleDispatch: mockToggleDispatch,
-    };
-
-    it('should not render mpu ads on AMP when ampAds toggle is set to false', () => {
-      const { container } = render(
-        <ServiceContextProvider service="mundo">
-          <RequestContextProvider
-            bbcOrigin="https://www.test.bbc.co.uk"
-            id="c0000000000o"
-            isAmp
-            pageType="frontPage"
-            service="mundo"
-            statusCode={200}
-            pathname="/mundo"
-          >
-            <ToggleContext.Provider value={toggleContextMock}>
-              <AdContainer slotType="mpu" />
-            </ToggleContext.Provider>
-          </RequestContextProvider>
-        </ServiceContextProvider>,
-      );
-
-      expect(container.firstChild).toBeNull();
-    });
-
-    it('should not render leaderboard ads on AMP when ampAds toggle is set to false', () => {
-      const { container } = render(
-        <ServiceContextProvider service="mundo">
-          <RequestContextProvider
-            bbcOrigin="https://www.test.bbc.co.uk"
-            id="c0000000000o"
-            isAmp
-            pageType="frontPage"
-            service="mundo"
-            statusCode={200}
-            pathname="/mundo"
-          >
-            <ToggleContext.Provider value={toggleContextMock}>
-              <AdContainer slotType="leaderboard" />
-            </ToggleContext.Provider>
-          </RequestContextProvider>
-        </ServiceContextProvider>,
-      );
-
-      expect(container.firstChild).toBeNull();
     });
   });
 });
